@@ -1,57 +1,83 @@
 {
-  "name": "saleor-storefront",
-  "version": "0.0.0",
-  "private": true,
-  "workspaces": [
-    "apps/*",
-    "packages/*"
-  ],
-  "scripts": {
-    "build": "env-cmd --no-override -f .env env-cmd --no-override -f .env.local --fallback turbo run build --cache-dir=.turbo",
-    "build:saleor-app-checkout": "pnpm run build --filter=saleor-app-checkout...",
-    "build:storefront": "pnpm run build --filter=storefront...",
-    "start": "env-cmd --no-override -f .env env-cmd --no-override -f .env.local --fallback turbo run start --parallel --cache-dir=.turbo",
-    "start:saleor-app-checkout": "pnpm run start --filter=saleor-app-checkout...",
-    "start:storefront": "pnpm run start --filter=storefront...",
-    "generate": "env-cmd --no-override -f .env env-cmd --no-override -f .env.local --fallback turbo run generate --cache-dir=.turbo",
-    "dev": "env-cmd --no-override -f .env env-cmd --no-override -f .env.local --fallback pnpm run _dev",
-    "/*": "https://github.com/vercel/turborepo/issues/460",
-    "_dev": "turbo run build --filter=@saleor/checkout-storefront --cache-dir=.turbo && turbo run dev --parallel --cache-dir=.turbo --filter=!checkout",
-    "lint": "env-cmd --no-override -f .env env-cmd --no-override -f .env.local --fallback turbo run lint --cache-dir=.turbo",
-    "lint:staged": "env-cmd --no-override -f .env env-cmd --no-override -f .env.local --fallback turbo run lint:staged",
-    "check-types": "turbo run check-types --cache-dir=.turbo",
-    "test": "env-cmd --no-override -f .env turbo run test --cache-dir=.turbo",
-    "i18n": "pnpm run i18n:extract && pnpm run i18n:compile",
-    "i18n:extract": "turbo run i18n:extract --cache-dir=.turbo",
-    "i18n:compile": "turbo run i18n:compile --cache-dir=.turbo",
-    "format": "prettier --write .",
-    "prepare": "husky install",
-    "clean-deps": "turbo run clean && rm -rf node_modules **/node_modules pnpm-lock.yaml",
-    "clean-envs": "find . -maxdepth 3 -name \".env.local\" -type f -exec rm -i {} \\;",
-    "clean-auth-files": "find . -maxdepth 3 -name \".auth_token\" -exec rm -i {} \\;",
-    "clean": "rm -rf .turbo; turbo run clean",
-    "cy:run": "cypress run",
-    "cy:open": "cypress open"
-  },
-  "devDependencies": {
-    "cypress": "10.11.0",
-    "env-cmd": "10.1.0",
-    "husky": "8.0.1",
-    "lint-staged": "13.0.3",
-    "prettier": "2.7.1",
-    "turbo": "1.6.1",
-    "typescript": "4.8.4"
-  },
-  "engines": {
-    "npm": ">=7.0.0",
-    "node": ">=16.0.0",
-    "pnpm": ">=8"
-  },
-  "packageManager": "pnpm@8.6.6",
-  "pnpm": {
-    "overrides": {
-      "@urql/exchange-auth>@urql/core": "3.2.2",
-      "@urql/exchange-multipart-fetch>@urql/core": "3.2.2"
-    }
-  }
+	"name": "saleor-storefront",
+	"version": "0.1.0",
+	"private": true,
+	"type": "module",
+	"scripts": {
+		"dev": "next dev",
+		"build": "pnpm run generate && pnpm run i18n && next build",
+		"start": "next start",
+		"lint": "next lint --dir src --fix",
+		"generate": "graphql-codegen --config .graphqlrc.ts",
+		"prepare": "husky install",
+		"i18n:extract": "formatjs extract 'src/checkout/src/**/*.{js,jsx,ts,tsx}' --format transifex --id-interpolation-pattern '[folder]/[name]/[sha512:contenthash:base64:6]' --out-file src/checkout/content/locales/en-US.json",
+		"i18n:compile": "formatjs compile-folder --ast --format transifex src/checkout/content/locales src/checkout/content/compiled-locales",
+		"i18n": "pnpm run i18n:extract && pnpm run i18n:compile"
+	},
+	"dependencies": {
+		"@adyen/adyen-web": "5.51.0",
+		"@adyen/api-library": "14.2.0",
+		"@apollo/client": "3.8.4",
+		"@headlessui/react": "1.7.17",
+		"@saleor/auth-sdk": "0.12.0",
+		"clsx": "2.0.0",
+		"editorjs-html": "3.4.3",
+		"formik": "2.4.5",
+		"libphonenumber-js": "1.10.45",
+		"lodash-es": "4.17.21",
+		"lucide-react": "0.279.0",
+		"next": "13.5.3",
+		"query-string": "8.1.0",
+		"react": "18.2.0",
+		"react-dom": "18.2.0",
+		"react-error-boundary": "4.0.11",
+		"react-intl": "6.4.7",
+		"react-spinners": "0.13.8",
+		"react-test-renderer": "18.2.0",
+		"react-toastify": "9.1.3",
+		"ts-invariant": "0.10.3",
+		"url-join": "5.0.0",
+		"urql": "4.0.5",
+		"yup": "1.3.2",
+		"zustand": "4.4.1"
+	},
+	"devDependencies": {
+		"@formatjs/cli": "6.2.0",
+		"@graphql-codegen/cli": "5.0.0",
+		"@graphql-codegen/client-preset": "4.1.0",
+		"@graphql-typed-document-node/core": "3.2.0",
+		"@next/env": "13.5.3",
+		"@parcel/watcher": "2.3.0",
+		"@tailwindcss/forms": "0.5.6",
+		"@tailwindcss/typography": "0.5.10",
+		"@types/lodash-es": "4.17.9",
+		"@types/node": "20.8.0",
+		"@types/react": "18.2.24",
+		"@types/react-dom": "18.2.8",
+		"@types/url-join": "4.0.1",
+		"@typescript-eslint/eslint-plugin": "6.7.3",
+		"@typescript-eslint/parser": "6.7.3",
+		"autoprefixer": "10.4.16",
+		"eslint": "8.50.0",
+		"eslint-config-next": "13.5.3",
+		"eslint-config-prettier": "9.0.0",
+		"eslint-plugin-formatjs": "4.10.5",
+		"eslint-plugin-import": "2.28.1",
+		"graphql-tag": "2.12.6",
+		"husky": "8.0.3",
+		"lint-staged": "14.0.1",
+		"postcss": "8.4.31",
+		"prettier": "3.0.3",
+		"prettier-plugin-tailwindcss": "0.5.4",
+		"tailwindcss": "3.3.3",
+		"typescript": "5.2.2",
+		"wonka": "6.3.4"
+	},
+	"resolutions": {
+		"graphql": "16.8.1"
+	},
+	"engines": {
+		"pnpm": ">=8.8.0",
+		"node": "^20.0.0"
+	}
 }
