@@ -1,3 +1,11 @@
-module.exports = {
-  "*.{mjs,cjs,js,jsx,mts,cts,ts,tsx,css,md,json,yml,yaml}": "prettier --write",
+// https://nextjs.org/docs/basic-features/eslint#lint-staged
+
+import path from "path";
+
+const buildEslintCommand = (filenames) =>
+	`next lint --fix --file ${filenames.map((f) => path.relative(process.cwd(), f)).join(" --file ")}`;
+
+export default {
+	"*.{js,cjs,mjs,jsx,ts,cts,mts,tsx}": [buildEslintCommand],
+	"*.*": "prettier --write --ignore-unknown",
 };
